fix ndcg

+++ python/utils.py
@@ -207,8 +207,8 @@ def evaluate_model(model, dataset, args, mode='test'):
             sys.stdout.flush()
 
     return_vals = { 
-        'NDCG@10': int(NDCG / valid_user),
-        'NDCG@5': int(NDCG_5 / valid_user),
+        'NDCG@10': NDCG / valid_user,
+        'NDCG@5': NDCG_5 / valid_user,
         'HR@10': HT / valid_user,
         'HR@5': HT_5 / valid_user,
         'MRR': MRR / valid_user
